Support multiple phpunit version ArraySubset
