+++ phpstan.neon.dist
@@ -10,16 +10,19 @@ parameters:
 		constant_hassers: false
 	bootstrapFiles:
 		- vendor/bin/.phpunit/phpunit/vendor/autoload.php
-		- tests/Fixtures/app/AppKernel.php
 		# We're aliasing classes for phpunit in this file, it needs to be added here see phpstan/#2194
 		- src/Bridge/Symfony/Bundle/Test/Constraint/ArraySubset.php
+		- tests/Fixtures/app/AppKernel.php
 	excludes_analyse:
 		- tests/Fixtures/app/var/cache
 		# The Symfony Configuration API isn't good enough to be analysed
 		- src/Bridge/Symfony/Bundle/DependencyInjection/Configuration.php
+		# Phpstan runs on phpunit > 9, a signature changed in this file
+		- src/Bridge/Symfony/Bundle/Test/Constraint/ArraySubsetLegacy.php
 		# Imported code (temporary)
 		- src/Bridge/Symfony/Bundle/Test/BrowserKitAssertionsTrait.php
 		- tests/Bridge/Symfony/Bundle/Test/WebTestCaseTest.php
+		- tests/ProphecyTrait.php
 	earlyTerminatingMethodCalls:
 		PHPUnit\Framework\Constraint\Constraint:
 			- fail
