+++ src/Bridge/Symfony/Bundle/Test/Constraint/ArraySubset.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 
 namespace ApiPlatform\Core\Bridge\Symfony\Bundle\Test\Constraint;
 
-use PHPUnit\Framework\Constraint\Constraint;
 use PHPUnit\SebastianBergmann\Comparator\ComparisonFailure;
 use SebastianBergmann\Comparator\ComparisonFailure as LegacyComparisonFailure;
 
@@ -21,86 +20,9 @@ if (!class_exists(ComparisonFailure::class)) {
     class_alias(LegacyComparisonFailure::class, 'PHPUnit\SebastianBergmann\Comparator\ComparisonFailure');
 }
 
-/**
- * Constraint that asserts that the array it is evaluated for has a specified subset.
- *
- * Uses array_replace_recursive() to check if a key value subset is part of the
- * subject array.
- *
- * Imported from dms/phpunit-arraysubset-asserts, because the original constraint has been deprecated.
- *
- * @copyright Sebastian Bergmann <sebastian@phpunit.de>
- * @copyright Rafael Dohms <rdohms@gmail.com>
- *
- * @see https://github.com/sebastianbergmann/phpunit/issues/3494
- */
-final class ArraySubset extends Constraint
-{
-    private $subset;
-    private $strict;
-
-    public function __construct(iterable $subset, bool $strict = false)
-    {
-        $this->strict = $strict;
-        $this->subset = $subset;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function evaluate($other, string $description = '', bool $returnResult = false): ?bool
-    {
-        //type cast $other & $this->subset as an array to allow
-        //support in standard array functions.
-        $other = $this->toArray($other);
-        $this->subset = $this->toArray($this->subset);
-        $patched = array_replace_recursive($other, $this->subset);
-        if ($this->strict) {
-            $result = $other === $patched;
-        } else {
-            $result = $other == $patched;
-        }
-        if ($returnResult) {
-            return $result;
-        }
-        if ($result) {
-            return null;
-        }
-
-        $f = new ComparisonFailure(
-            $patched,
-            $other,
-            var_export($patched, true),
-            var_export($other, true)
-        );
-        $this->fail($other, $description, $f);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function toString(): string
-    {
-        return 'has the subset '.$this->exporter()->export($this->subset);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function failureDescription($other): string
-    {
-        return 'an array '.$this->toString();
-    }
-
-    private function toArray(iterable $other): array
-    {
-        if (\is_array($other)) {
-            return $other;
-        }
-        if ($other instanceof \ArrayObject) {
-            return $other->getArrayCopy();
-        }
-
-        return iterator_to_array($other);
-    }
+// Aliases as string to avoid loading the class
+if (\PHP_VERSION_ID >= 80000 || (float) getenv('SYMFONY_PHPUNIT_VERSION') > 8) {
+    class_alias('ApiPlatform\Core\Bridge\Symfony\Bundle\Test\Constraint\ArraySubsetV9', 'ApiPlatform\Core\Bridge\Symfony\Bundle\Test\Constraint\ArraySubset');
+} else {
+    class_alias('ApiPlatform\Core\Bridge\Symfony\Bundle\Test\Constraint\ArraySubsetLegacy', 'ApiPlatform\Core\Bridge\Symfony\Bundle\Test\Constraint\ArraySubset');
 }
